% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELLiD.R
\name{CELLiDCluster}
\alias{CELLiDCluster}
\title{Cluster-level cell type prediction}
\usage{
CELLiDCluster(
  rna,
  ref_data = NULL,
  ref_deg = NULL,
  atlas = NULL,
  n_predict = 1,
  verbose = T,
  ref_path = NULL,
  ncores = 5
)
}
\arguments{
\item{rna}{A data.frame or matrix with genes listed as row names}

\item{ref_data}{Reference gene expression. If not specified, reference gene
expression data will be downloaded from the DISCO database.}

\item{ref_deg}{DEG list.}

\item{atlas}{Specify the atlas. If not specified, CELLiD will use all data.}

\item{n_predict}{Number of predicted cell type. Maximum value is 3}

\item{ref_path}{Specify the file path to save the downloaded reference data.}

\item{ncores}{Number of core used for CELLiD}
}
\description{
Cluster-level cell type prediction
}
