% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELLiD.R
\name{CELLiDEnrichment}
\alias{CELLiDEnrichment}
\title{Geneset enrichment analysis using DISCO data}
\usage{
CELLiDEnrichment(input, reference = NULL, ref.path = NULL, ncores = 10)
}
\arguments{
\item{input}{A one or two columns data.frame. The first column is gene and the second column is log2FC (optional)}

\item{reference}{Reference geneset If not specified, reference geneset will be downloaded from the DISCO database.}

\item{ref.path}{Specify the file path to save the downloaded reference geneset data.}

\item{ncores}{Number of core used for CELLiD}
}
\description{
Geneset enrichment analysis using DISCO data
}
