% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadDiscoData.R
\name{DownloadDiscoData}
\alias{DownloadDiscoData}
\title{Download data from DISCO database using the filtered metadata}
\usage{
DownloadDiscoData(metadata, output_dir = "DISCOtmp")
}
\arguments{
\item{metadata}{The resulting metadata object produced by FilterDiscoMetadata function}

\item{output_dir}{The directory where the downloaded files will be stored}
}
\description{
Download data from DISCO database using the filtered metadata
}
\examples{
# Download the data from project 'GSE174748' and store it in the 'disco_data' directory
metadata = FilterDiscoMetadata(
  project = "GSE174748"
)
DownloadDiscoData(metadata, output_dir = "disco_data")
}
