% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetadata.R
\name{GetCellTypeChildren}
\alias{GetCellTypeChildren}
\title{Check children of input cell type}
\usage{
GetCellTypeChildren(cell_type, cell_ontology = NULL)
}
\arguments{
\item{cell_type}{The cell type of the input, which may be a string or a list of string}

\item{cell_ontology}{The cell type ontology. If not specified, the cell type ontology will be retrieved from the DISCO database.}
}
\value{
List of cell type
}
\description{
Check children of input cell type
}
\examples{
# Get children of B cell
GetCellTypeChildren(cell_type = c("B cell"))

# Get children of B cell and Macrophage
GetCellTypeChildren(cell_type = c("B cell", "Macrophage"))
}
