% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetadata.R
\name{GetSampleMetadata}
\alias{GetSampleMetadata}
\title{Cluster-level cell type prediction}
\usage{
GetSampleMetadata(
  project = NULL,
  tissue = NULL,
  disease = NULL,
  sample.type = NULL,
  cell.type = NULL
)
}
\arguments{
\item{rna}{A data.frame or matrix with genes listed as row names}

\item{ref.data}{Reference gene expression. If not specified, reference gene
expression data will be downloaded from the DISCO database.}

\item{ref.deg}{DEG list.}

\item{atlas}{Specify the atlas. If not specified, CELLiD will use all data.}

\item{n.predict}{Number of predicted cell type. Maximum value is 3}

\item{ref.path}{Specify the file path to save the downloaded reference data.}

\item{ncores}{Number of core used for CELLiD}
}
\description{
Cluster-level cell type prediction
}
