% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadDiscoData.R
\name{DownloadDiscoData}
\alias{DownloadDiscoData}
\title{Check children of input cell type}
\usage{
DownloadDiscoData(metadata, output.dir = "DISCOtmp")
}
\arguments{
\item{cell.type}{The cell type of the input, which may be a string or a list of string}

\item{cell.ontology}{The cell type ontology. If not specified, the cell type ontology will be retrieved from the DISCO database.}
}
\value{
List of cell type
}
\description{
Check children of input cell type
}
\examples{
# Get children of B cell
GetCellTypeChildren(cell.type = c("B cell"))

# Get children of B cell and Macrophage
GetCellTypeChildren(cell.type = c("B cell", "Macrophage"))
}
