% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetadata.R
\name{FilterDiscoMetadata}
\alias{FilterDiscoMetadata}
\title{Cluster-level cell type prediction}
\usage{
FilterDiscoMetadata(
  project = NULL,
  tissue = NULL,
  disease = NULL,
  sample.type = NULL,
  cell.type = NULL,
  include.cell.type.children = T,
  min.cell.per.sample = 100
)
}
\arguments{
\item{project}{project ID(s) of interest, can be a string or vector}

\item{tissue}{tissue(s) of interest, can be a string or vector}

\item{disease}{disease(s) of interest, can be a string or vector}

\item{sample.type}{sample type(s) of interest, can be a string or vector}

\item{cell.type}{cell type(s) of interest, can be a string or vector}

\item{include.cell.type.children}{Whether the cell type is an exact match or includes its child elements}

\item{min.cell.per.sample}{Only samples with a cell count greater than the minimum cell count per sample will be retained.}
}
\description{
Cluster-level cell type prediction
}
