% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetadata.R
\name{FilterDiscoMetadata}
\alias{FilterDiscoMetadata}
\title{Get metadata from DISCO and apply filters on it}
\usage{
FilterDiscoMetadata(
  sample_id = NULL,
  project_id = NULL,
  tissue = NULL,
  disease = NULL,
  platform = NULL,
  sample_type = NULL,
  cell_type = NULL,
  cell_type_confidence = "medium",
  include_cell_type_children = T,
  min_cell_per_sample = 100
)
}
\arguments{
\item{tissue}{tissue(s) of interest, can be a string or vector}

\item{disease}{disease(s) of interest, can be a string or vector}

\item{platform}{platform(s) of interest, can be a string or vector}

\item{sample_type}{sample type(s) of interest, can be a string or vector}

\item{cell_type}{cell type(s) of interest, can be a string or vector}

\item{cell_type_confidence}{Filter the results based on the confidence of cell type prediction, which can be categorized as high, medium, or all}

\item{include_cell_type_children}{Whether the cell type is an exact match or includes its child elements}

\item{min_cell_per_sample}{Only samples with a cell count greater than the minimum cell count per sample will be retained.}

\item{sample}{sample ID(s) of interest, can be a string or vector}

\item{project}{project ID(s) of interest, can be a string or vector}
}
\description{
Get metadata from DISCO and apply filters on it
}
